Move player functions to player.c
